+++ cuiman/tests/cli/test_client.py
@@ -7,9 +7,6 @@ from unittest import TestCase
 import pytest
 import typer
 
-# noinspection PyProtectedMember
-from typer._click import exceptions as click_exceptions
-
 from cuiman import ClientError
 from cuiman.api.client import Client
 from cuiman.api.transport import TransportError
@@ -29,7 +26,7 @@ class UseClientTest(TestCase):
 
     # noinspection PyMethodMayBeStatic
     def test_fail_with_client_error(self):
-        with pytest.raises(click_exceptions.Exit):
+        with pytest.raises(typer.Exit):
             with use_client(new_cli_context(), None):
                 raise ClientError(
                     "Not found",
@@ -59,7 +56,7 @@ class UseClientTest(TestCase):
 
     # noinspection PyMethodMayBeStatic
     def test_fail_with_transport_error(self):
-        with pytest.raises(click_exceptions.Exit):
+        with pytest.raises(typer.Exit):
             with use_client(new_cli_context(), None):
                 raise TransportError("connection could not be established")